rename tests to codegen
