+++ tests/test_codegen/schemas/object_with_list_of_type_with_union.py
