+++ tests/test_codegen/schemas/object_with_object.py
