+++ tests/test_codegen/schemas/object_with_optional_object.py
