+++ tests/test_codegen/schemas/object_with_optional_scalar.py
