+++ tests/test_codegen/schemas/object_with_scalar.py
