+++ tests/test_codegen/schemas/object_with_union.py
