+++ tests/test_codegen/test_py/test_simple_class.py
@@ -3,7 +3,7 @@ from qtgql.codegen.objecttype import FieldProperty, GqlType, Kinds
 from qtgql.codegen.py.compiler import SchemaTemplate
 from qtgql.typingref import TypeHinter
 
-from tests.test_compiler.test_py.conftest import generate_type_kwargs, property_tester
+from tests.test_codegen.test_py.conftest import generate_type_kwargs, property_tester
 
 
 @pytest.fixture
